+++ src/omnifuse/__init__.py
@@ -3,7 +3,7 @@
 from importlib import import_module
 
 
-__version__ = "0.8.0"
+__version__ = "0.8.1"
 
 _EXPORTS = {
     "KnowledgeSearch": (".knowledge_search", "KnowledgeSearch"),